Create Roles list and the plot
You are given a list of roles and a script for the play as an array of lines.

Each line of the play script is presented as follows:
Role: text

The text can contain any characters.

Write a method that will group lines by role, number them, and return the result in the form of ready-made text (see example). Each group is printed as follows:

Role:
i) text
j) text2
...
== line feed ==

i and j are line numbers in the script. Row indexing begins with one, groups should be displayed in accordance with the order of roles. Line breaks between groups are required, line breaks at the end of the text are not taken into account.

Note that you have to process a huge piece of 50,000 lines for 10 roles - accordingly, incorrect assembly of the resulting line may exceed the time limit.

Pay attention to a few more nuances:

a character's name can appear more than once in a line, including a colon;
the name of one role can be a prefix of the name of another role (for example, "Luka" and "Luka Lukic");
a role that does not have replicas must also be present in the output file;
use the '\ n' character (UNIX style line feed) as a line feed;
be careful not to add extra spaces at the end of lines.