Create AsciiCharSequence
Write an AsciiCharSequence class that implements compact storage of a sequence of ASCII characters (their codes fit into one byte) in a byte array. Compared to the String class, which stores each character as a char, AsciiCharSequence will consume half the memory.

The AsciiCharSequence class must:

implement the java.lang.CharSequence interface;
have a constructor that takes an array of bytes;
define methods length (), charAt (), subSequence () and toString ()
See the java.lang.CharSequence interface (JavaDoc or sources) for method signatures and expectations for their behavior.

In this task, the charAt () and subSequence () methods will always be supplied with correct input parameters, so there is no need to check them and handle errors. Moreover, we have not yet passed the exception.