Create getCallerClassAndMethodName
Implement a method that lets other methods know where they were called from.

The getCallerClassAndMethodName () method must return the name of the class and method from which the method that called this utility method was called. Or null (a null reference, not the string "null") if the method that called getCallerClassAndMethodName () is the entry point to the program, i.e. no one called him.

This is relevant, for example, in logging libraries, where for each message in the log it is necessary to display the class and method from which the message was logged.