Parentheses in code
You are developing a text editor for
programmers and want to implement verification
the correctness of the brackets. In code can
there are brackets [] {} (). Of these, brackets [, {
and (are considered opening, and their corresponding closing parentheses are
],} and).
In case the brackets are placed incorrectly, the editor should
also tell the user the first place where the error is found.
First of all, you need to find a closing parenthesis, for which either there is no corresponding opening (for example, parenthesis] in
the line “] ()”), or it closes the open parenthesis that does not match it (example: “() [}”). If there are no such errors, it is necessary to find the first opening parenthesis for which there is no corresponding closing one (example: parenthesis (in the line “{} ([]”)).
In addition to parentheses, the source code can contain Latin characters, numbers and punctuation marks.
Input format. String s [1. ... ... n], consisting of capital and uppercase letters of the Latin alphabet, numbers, punctuation marks and brackets from the set [] {} ().
Output format. If the parentheses in s are placed correctly, print
the string “Success". Otherwise, print the index (using indexing from one) of the first closing parenthesis, for
which has no corresponding opening. If not,
print the index of the first opening parenthesis for which there is no
the corresponding closing.