Create Tree height
Trees have a wide variety of uses in Computer Science. They are used
both for data representation and in many machine learning algorithms. Further
we will also learn how balanced trees are used to implement vocabularies
and associative arrays. Structure data
data are used in one way or another in all
programming languages ​​and databases.
Your goal in this task is to learn how to store and efficiently process trees, even if they have hundreds of thousands of vertices.
Input format. The first line contains a natural number n. The second
string contains n integers parent0
,. ... ... , parentn − 1
... For each 0 ≤ i ≤ n − 1, parenti is the parent of vertex i; if parenti = −1,
then i is a root. It is guaranteed that there is exactly one root. It is guaranteed that the given sequence defines a tree.
Output format. The height of the tree.
Limitations. 1 ≤ n ≤ 105