Chaining hashing
Chaining hashing is one of the most popular implementation methods
hash tables in practice. Your goal in this problem is to implement such a scheme using a table with m cells and a polynomial hash function on rows
h (S) =


|
X
S | −1
i = 0
S [i] x
i mod p

 mod m,
where S [i] - ASCII-code of the i-th character of string S, p = 1,000,000 007 - simple
number, and x = 263. Your program should support the following
request types:
• add string: add a string to the table. If such
the line is already there, ignore the request;
• del string: remove the string string from the table. If such
no line, ignore request;
• find string: print "yes" or "no" depending on whether
is the string string in the table or not;
• check i: print the i-th list (using a space as a separator); if the i-th list is empty, output an empty string.
When adding a string to a chain, the string must be added to the beginning of the chain.