Parallel processing
Given n processors and m tasks, determine, for each task,
what processor it will be processed by.
Entrance. Number of processors n and sequence of numbers
t0,. ... ... , tm − 1, where ti is the time required to process the ith
tasks.
Exit. For each task, determine which processor
and at what time will it start processing it, assuming that
each task goes to the first free processor for processing.
In this task, your goal is to simulate parallel task list processing. Such handlers (dispatchers) are in
all operating systems.
You have n processors and a sequence of m tasks. For
each task is given the time required to process it. The next job goes to the first available processor (that is, if
there are several available processors, then the available processor with the minimum number gets this job).