Building the heap
Building the heap is a key step in the heap sort algorithm. This algorithm has a worst-case O (n log n) runtime, unlike the quicksort algorithm, which guarantees such an estimate only in the average. The quicksort algorithm is more often used in practice, since in most
cases it works faster, but the algorithm
heap sort is used for external sorting of data when needed
sort out huge data,
not fit into the computer's memory.
To turn this array into a heap, you need to do
several exchanges of its elements. We call the basic exchange
an operation that swaps the elements A [i] and A [j]. Your goal
in this task - convert a given array into a heap in a linear number of exchanges.